requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
docker==6.1.3
Flask==2.3.2
Flask-MySQLdb==1.0.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.0.33
mysqlclient==2.1.1
packaging==23.1
protobuf==3.20.3
pywin32==306
requests==2.31.0
typing_extensions==4.6.3
urllib3==2.0.2
websocket-client==1.5.2
Werkzeug==2.3.4
